Fix problem on 32bit devices when upgrading
Backport of 32bf01cde3203

+++ src/realm/cluster.cpp
@@ -340,13 +340,13 @@ ref_type ClusterNodeInner::insert(ObjKey key, const FieldValues& init_values, Cl
         size_t new_ref_ndx = child_info.ndx + 1;
 
         int64_t split_key_value = state.split_key + child_info.offset;
-        size_t sz = node_size();
+        uint64_t sz = node_size();
         if (sz < cluster_node_size) {
             if (m_keys.is_attached()) {
                 m_keys.insert(new_ref_ndx, split_key_value);
             }
             else {
-                if (size_t(split_key_value) != sz << m_shift_factor) {
+                if (uint64_t(split_key_value) != sz << m_shift_factor) {
                     ensure_general_form();
                     m_keys.insert(new_ref_ndx, split_key_value);
                 }
@@ -592,7 +592,7 @@ void ClusterNodeInner::add(ref_type ref, int64_t key_value)
         m_keys.add(key_value);
     }
     else {
-        if (size_t(key_value) != (node_size() << m_shift_factor)) {
+        if (uint64_t(key_value) != (uint64_t(node_size()) << m_shift_factor)) {
             ensure_general_form();
             m_keys.add(key_value);
         }
